+++ webarchive/exceptions.py
@@ -1,0 +1,10 @@
+"""Webarchive-related exceptions."""
+
+
+__all__ = ["WebArchiveError"]
+
+
+class WebArchiveError(Exception):
+    """Class representing an error in an operation on a WebArchive."""
+
+    pass